+++ src/Data/TimerWheel.hs
@@ -1,13 +1,10 @@
-{-# LANGUAGE CPP, LambdaCase, MagicHash, NamedFieldPuns, RecursiveDo,
-             ScopedTypeVariables, TypeApplications, UnboxedTuples,
-             ViewPatterns #-}
-
-{-# options_ghc -funbox-strict-fields #-}
+{-# OPTIONS_GHC -funbox-strict-fields #-}
 
 module Data.TimerWheel
   ( -- * Timer wheel
     TimerWheel
   , new
+  , Config(..)
   , register
   , register_
   , recurring
@@ -17,37 +14,29 @@ module Data.TimerWheel
 
 import Entries (Entries)
 import Supply  (Supply)
+import Utils
+import Wheel   (Wheel)
 
-import qualified Entries as Entries
+import qualified Entries
 import qualified Supply
+import qualified Wheel
 
-import Control.Concurrent           (forkIOWithUnmask, myThreadId, threadDelay,
-                                     throwTo)
+import Control.Concurrent      (forkIOWithUnmask, myThreadId, throwTo)
 import Control.Concurrent.MVar
-import Control.Exception            (Exception(fromException, toException),
-                                     SomeException, asyncExceptionFromException,
-                                     asyncExceptionToException, catch, mask_,
-                                     throwIO)
-import Control.Monad                (join, void, when)
-import Data.Coerce                  (coerce)
-import Data.Fixed                   (E6, Fixed(MkFixed))
-import Data.Foldable                (for_)
-import Data.IORef                   (IORef, newIORef, readIORef, writeIORef)
-import Data.Primitive.MutVar        (MutVar, atomicModifyMutVar', newMutVar)
-import Data.Primitive.UnliftedArray (MutableUnliftedArray,
-                                     UnliftedArray(UnliftedArray),
-                                     freezeUnliftedArray, indexUnliftedArray,
-                                     sizeofUnliftedArray,
-                                     unsafeNewUnliftedArray, writeUnliftedArray)
-import Data.Word                    (Word64)
-import GHC.Base                     (IO(IO), mkWeak#)
-#if MIN_VERSION_base(4,11,0)
-import GHC.Clock (getMonotonicTimeNSec)
-#else
-import System.Clock (Clock(Monotonic), getTime, toNanoSecs)
-#endif
-import GHC.Prim         (RealWorld)
-import GHC.Weak         (Weak(Weak), deRefWeak)
+import Control.Exception       (Exception(fromException, toException),
+                                SomeException, asyncExceptionFromException,
+                                asyncExceptionToException, catch, mask_,
+                                throwIO)
+import Control.Monad           (join, void, when)
+import Data.Coerce             (coerce)
+import Data.Fixed              (E6, Fixed(MkFixed))
+import Data.IORef              (IORef, newIORef, readIORef, writeIORef)
+import Data.Primitive.MutVar   (MutVar, atomicModifyMutVar')
+import Data.Word               (Word64)
+-- import GHC.Base                     (IO(IO), mkWeak#)
+import GHC.Generics (Generic)
+import GHC.Prim     (RealWorld)
+-- import GHC.Weak         (Weak(Weak), deRefWeak)
 import Numeric.Natural  (Natural)
 import System.IO.Unsafe (unsafeInterleaveIO)
 
@@ -109,13 +98,18 @@ import System.IO.Unsafe (unsafeInterleaveIO)
 --    +-----+-----+-----+-----+-----+-----+-----+-----+
 -- @
 data TimerWheel =  TimerWheel
-  { wheelResolution :: !Word64
-    -- ^ The length of time that each entry corresponds to, in nanoseconds.
-  , wheelSupply     :: !Supply
+  { -- ^ The length of time that each entry corresponds to, in nanoseconds.
+    wheelSupply :: !Supply
     -- ^ A supply of unique ints.
-  , wheelEntries    :: !(UnliftedArray (MutVar RealWorld Entries))
+  , wheelWheel :: !Wheel
     -- ^ The array of collections of timers.
   }
+
+data Config
+  = Config
+  { spokes :: !Natural
+  , resolution :: !(Fixed E6)
+  } deriving stock (Generic, Show)
 
 newtype TimerWheelDied
   = TimerWheelDied SomeException
@@ -126,7 +120,7 @@ instance Exception TimerWheelDied where
   fromException = asyncExceptionFromException
 
 data InvalidTimerWheelConfig
-  = InvalidTimerWheelConfig !Natural !(Fixed E6)
+  = InvalidTimerWheelConfig !Config
   deriving (Show)
 
 instance Exception InvalidTimerWheelConfig
@@ -134,133 +128,45 @@ instance Exception InvalidTimerWheelConfig
 -- | @new n s@ creates a 'TimerWheel' with __@n@__ spokes and a resolution of
 -- __@s@__ seconds.
 --
--- /Throws./ If @n <= 0@, @n > maxBound \@Int@ or @s@ are @<= 0@, throws an
+-- /Throws./ If @n == 0@, @n > maxBound \@Int@ or @s <= 0@, throws an
 -- 'InvalidTimerWheelConfig' exception.
 --
 -- /Throws./ If the timeout thread dies, asynchronously throws 'TimerWheelDied'
 -- to the thread that called 'new'.
-new :: Natural -> Fixed E6 -> IO TimerWheel
-new slots resolution = do
-  when (invalidConfig slots resolution)
-    (throwIO (InvalidTimerWheelConfig slots resolution))
+new :: Config -> IO TimerWheel
+new config@(Config { spokes, resolution }) = do
+  when (invalidConfig config)
+    (throwIO (InvalidTimerWheelConfig config))
 
-  wheel :: UnliftedArray (MutVar RealWorld Entries) <-
-    newWheel slots
+  wheel :: Wheel <-
+    Wheel.new
+      (fromIntegral spokes)
+      (fromIntegral (coerce resolution :: Integer))
 
   supply :: Supply <-
     Supply.new
 
-  weakWheel :: Weak (UnliftedArray (MutVar RealWorld Entries)) <-
-    weakUnliftedArray wheel
-
   do
     thread <- myThreadId
     void $ mask_ $ forkIOWithUnmask $ \unmask ->
-      unmask (reaper resolution (sizeofUnliftedArray wheel) weakWheel)
+      unmask (Wheel.reap wheel)
         `catch` \e -> throwTo thread (TimerWheelDied e)
 
   pure TimerWheel
-    { wheelResolution = microToNano resolution
-    , wheelSupply = supply
-    , wheelEntries = wheel
+    { wheelSupply = supply
+    , wheelWheel = wheel
     }
 
-invalidConfig :: Natural -> Fixed E6 -> Bool
-invalidConfig slots resolution =
+invalidConfig :: Config -> Bool
+invalidConfig Config { spokes, resolution } =
   or
-    [ slots <= 0
-    , slots > fromIntegral (maxBound :: Int)
+    [ spokes == 0
+    , spokes > fromIntegral (maxBound :: Int)
     , resolution <= 0
     ]
 
--- Initialize a wheel with the given number of slots.
---
--- Precondition: 0 < slots <= maxBound @Int
-newWheel :: Natural -> IO (UnliftedArray (MutVar RealWorld Entries))
-newWheel (fromIntegral -> slots) = do
-  wheel :: MutableUnliftedArray RealWorld (MutVar RealWorld Entries) <-
-    unsafeNewUnliftedArray slots
-  for_ [0..slots-1] $ \i ->
-    writeUnliftedArray wheel i =<< newMutVar Entries.empty
-  freezeUnliftedArray wheel 0 slots
-
-weakUnliftedArray :: UnliftedArray a -> IO (Weak (UnliftedArray a))
-weakUnliftedArray array =
-  case array of
-    UnliftedArray array# ->
-      IO $ \s ->
-        case mkWeak# array# array (\t -> (# t, () #)) s of
-          (# s', x #) ->
-            (# s', Weak x #)
-
-reaper
-  :: Fixed E6
-  -> Int
-  -> Weak (UnliftedArray (MutVar RealWorld Entries))
-  -> IO ()
-reaper (MkFixed micro) len weakWheel =
-  loop 0
-
-  where
-    -- Reaper loop: we haven't yet run the entries at index 'i'.
-    loop :: Int -> IO ()
-    loop i = do
-      -- Sleep until the roughly the next bucket.
-      threadDelay (fromIntegral @Integer (coerce micro))
-
-      now :: Word64 <-
-        getMonotonicTime
-
-      -- De-ref the wheel, and if it's gone, we go too.
-      deRefWeak weakWheel >>= \case
-        Nothing ->
-          pure ()
-
-        Just wheel -> do
-          -- Figure out which bucket we're in. Usually this will be 'i+1', but
-          -- maybe we were scheduled a bit early and ended up in 'i', or maybe
-          -- running the entries in bucket 'i-1' took a long time and we slept
-          -- all the way to 'i+2'. In any case, we should run the entries in
-          -- buckets up-to-but-not-including this one, beginning with bucket
-          -- 'i'.
-
-          let
-            j :: Int
-            j =
-              fromIntegral (now `div` microToNano (MkFixed micro))
-                `mod` len
-
-          let
-            is :: [Int]
-            is =
-              if j >= i
-                then [i .. j-1]
-                else [i .. len - 1] ++ [0 .. j-1]
-
-          -- To actually run the entries in a bucket, partition them into
-          -- expired (count == 0) and alive (count > 0). Run the expired entries
-          -- and decrement the alive entries' counts by 1.
-
-          for_ is $ \k -> do
-            let
-              entriesRef :: MutVar RealWorld Entries
-              entriesRef =
-                indexUnliftedArray wheel k
-
-            join
-              (atomicModifyMutVar' entriesRef
-                (\entries ->
-                  if Entries.null entries
-                    then
-                      (entries, pure ())
-                    else
-                      case Entries.partition entries of
-                        (expired, alive) ->
-                          (alive, sequence_ expired)))
-          loop j
-
--- | @register n m w@ registers an action __@m@__ in timer wheel __@w@__ to fire
--- after __@n@__ seconds.
+-- | @register wheel action delay@ registers an action __@action@__ in timer
+-- wheel __@wheel@__ to fire after __@delay@__ seconds.
 --
 -- Returns an action that, when called, attempts to cancel the timer, and
 -- returns whether or not it was successful (@False@ means the timer has already
@@ -268,17 +174,22 @@ reaper (MkFixed micro) len weakWheel =
 --
 -- Subsequent calls to the cancel action have no effect, and continue to return
 -- whatever the first result was.
-register :: Fixed E6 -> IO () -> TimerWheel -> IO (IO Bool)
-register (microToNano -> delay) action wheel = do
+register :: TimerWheel -> IO () -> Fixed E6 -> IO (IO Bool)
+register wheel action (MkFixed (fromIntegral -> delay)) = do
   newEntryId :: Int <-
     Supply.next (wheelSupply wheel)
 
-  entriesVar :: MutVar RealWorld Entries <-
-    entriesIn delay wheel
+  now :: Word64 <-
+    getMonotonicMicros
 
-  atomicModifyMutVar' entriesVar
+  let
+    bucketVar :: MutVar RealWorld Entries
+    bucketVar =
+      Wheel.bucket (wheelWheel wheel) now
+
+  atomicModifyMutVar' bucketVar
     (\entries ->
-      (Entries.insert newEntryId entryCount action entries, ()))
+      (Entries.insert newEntryId count (const action entries) entries, ()))
 
   canceledVar :: MVar (Maybe Bool) <-
     unsafeInterleaveIO (newMVar Nothing)
@@ -287,36 +198,33 @@ register (microToNano -> delay) action wheel = do
     modifyMVar canceledVar $ \result -> do
       canceled <-
         maybe
-          (atomicModifyMutVar' entriesVar
+          (atomicModifyMutVar' bucketVar
             (\entries ->
-              case Entries.delete newEntryId entries of
-                Nothing ->
-                  (entries, False)
-                Just entries' ->
-                  (entries', True)))
+              maybe
+                (entries, False)
+                (, True)
+                (Entries.delete newEntryId entries)))
           pure
           result
       pure (Just canceled, canceled)
 
   where
-    entryCount :: Word64
-    entryCount =
-      delay `div`
-        (fromIntegral (sizeofUnliftedArray (wheelEntries wheel))
-          * wheelResolution wheel)
+    count :: Word64
+    count =
+      delay `div` Wheel.lenMicros (wheelWheel wheel)
 
 -- | Like 'register', but for when you don't intend to cancel the timer.
-register_ :: Fixed E6 -> IO () -> TimerWheel -> IO ()
-register_ delay action wheel =
-  void (register delay action wheel)
+register_ :: TimerWheel -> IO () -> Fixed E6 -> IO ()
+register_ wheel action delay =
+  void (register wheel action delay)
 
--- | @recurring n m w@ registers an action __@m@__ in timer wheel __@w@__ to
--- fire every __@n@__ seconds.
+-- | @recurring wheel delay action@ registers an action __@action@__ in timer
+-- wheel __@wheel@__ to fire every __@delay@__ seconds.
 --
 -- Returns an action that, when called, cancels the recurring timer.
-recurring :: Fixed E6 -> IO () -> TimerWheel -> IO (IO ())
-recurring delay action wheel = mdo
-  cancel <- register delay (action' cancelRef) wheel
+recurring :: TimerWheel -> IO () -> Fixed E6 -> IO (IO ())
+recurring wheel action delay = mdo
+  cancel <- register wheel (action' cancelRef) delay
   cancelRef <- newIORef cancel
   pure (untilTrue (join (readIORef cancelRef)))
 
@@ -324,21 +232,8 @@ recurring delay action wheel = mdo
     action' :: IORef (IO Bool) -> IO ()
     action' cancelRef = do
       action
-      cancel <- register delay (action' cancelRef) wheel
+      cancel <- register wheel (action' cancelRef) delay
       writeIORef cancelRef cancel
-
--- @entriesIn delay wheel@ returns the bucket in @wheel@ that corresponds to
--- @delay@ nanoseconds from now.
-entriesIn :: Word64 -> TimerWheel -> IO (MutVar RealWorld Entries)
-entriesIn delay TimerWheel{wheelResolution, wheelEntries} = do
-  now <- getMonotonicTime
-  pure (index ((now+delay) `div` wheelResolution))
-
-  where
-    index :: Word64 -> MutVar RealWorld Entries
-    index i =
-      indexUnliftedArray wheelEntries
-        (fromIntegral i `rem` sizeofUnliftedArray wheelEntries)
 
 -- Repeat an IO action until it returns 'True'.
 untilTrue :: IO Bool -> IO ()
@@ -346,15 +241,3 @@ untilTrue action =
   action >>= \case
     True -> pure ()
     False -> untilTrue action
-
-microToNano :: Fixed E6 -> Word64
-microToNano (MkFixed micro) =
-  fromIntegral micro * 1000
-
-getMonotonicTime :: IO Word64
-getMonotonicTime =
-#if MIN_VERSION_base(4,11,0)
-  getMonotonicTimeNSec
-#else
-  fromIntegral . toNanoSecs <$> getTime Monotonic
-#endif