+++ CHANGELOG.md
@@ -6,7 +6,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to the [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
 ## Unreleased
-
 ### Changed
 - If the timer wheel reaper thread crashes, it will propagate the exception to
 the thread that spawned it.
@@ -14,6 +13,8 @@ the thread that spawned it.
 - The cancel action returned by `register` is now memoized, which fixes a bug
 involving trying to cancel a `recurring` timer twice. The second call used to
 spin forever and peg a CPU.
+- Use `Config` type for creating a timer wheel
+- Change argument order around
 
 ## [0.1.0] - 2018-07-18